feat(erasure): a pending retain must say so, carry a date, and be on a list
Counsel approved shipping the address retain rules ahead of the sweep only on
three conditions, because a bounded retain that nothing enforces becomes an
unbounded retain — the blanket exclusion refused in August, relabelled. The
rule itself is unchanged; these sit on top of it.

1. The rules no longer read as implemented. Every one carries
   `enforcementStatus: 'pending'`, a new optional field on `ErasureRule`, so
   the manifest and anything rendering from it can tell a recorded DECISION
   from a shipped behaviour.

2. `enforcementDeadline: '2027-02-01'`, and the gate FAILS once it passes. Two
   quarters, set by review discipline rather than by first breach: the sweep
   needs a purge marker on `inspections` (the agreement pass keys on `signedAt`
   + `purged_at IS NULL` and there is no equivalent) and a decision about which
   column starts an inspection's clock — a schema change and a migration, not a
   patch. Deliberately NOT derived from when the first address falls due, which
   is not computable until that clock column exists. Moving the date is allowed;
   moving it silently is what this prevents.

3. The gate refuses NEW unenforced retain rules. `PENDING_ENFORCEMENT` is a
   checked-in list of the ten rules allowed to be pending, checked BOTH ways —
   a pending rule missing from it fails, and a stale entry whose rule is gone
   fails too, so the list cannot decay into a blanket permit. For a bounded
   retain the DEFAULT is refusal: a `retain` that declares a `retention` and no
   `enforcementStatus` fails, so "unenforced" is never what happens when nobody
   says anything. The two swept signature_base64 rules are marked 'enforced'.

`ERASURE_OUT_OF_SCOPE` moves to `erasure-out-of-scope.ts`. The manifest was at
389 of 400 lines and this note would have pushed it over; extracting beats
compacting, because the thing that would have been compacted is the reasoning.
The gate now concatenates both sources before parsing either, so splitting the
register cannot halve what it sees — two lines, and both `arrayBody` calls work
unchanged. Manifest 289, register 157.

Also records at the top of the manifest that two of its justifications were
checked against the code on 2026-08-07 and found false — `reports.title`
("a human writes"; it is machine-written) and `created_by_ref` (an "opaque id"
holding an email). Both rules survived review because the reasoning read well.
The warning is the general form: a premise stated in a comment is not evidence,
so read what writes the column before relying on the paragraph.

The `reports.title` rationale itself is untouched — its amendment history is
being written separately so the two changes stay reviewable apart.
